# openapi.yaml
openapi: 3.1.0
info:
  title: Simple CRUD API
  description: Inventory and Reservation management system API
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
tags:
- name: Products
  description: Product management endpoints
- name: Reservations
  description: Reservation management endpoints
paths:
  /api/reservations:
    post:
      tags:
      - Reservations
      summary: Create a reservation for a product
      operationId: createReservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReservationResponse"
  /api/reservations/{id}/confirm:
    post:
      tags:
      - Reservations
      summary: Confirm a reservation by ID
      operationId: confirmReservation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReservationResponse"
  /api/products:
    post:
      tags:
      - Products
      summary: Create a new product
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /api/reservations/{id}:
    get:
      tags:
      - Reservations
      summary: Get reservation by ID
      operationId: getReservation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReservationResponse"
  /api/products/{id}:
    get:
      tags:
      - Products
      summary: Get product by ID
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
components:
  schemas:
    CreateReservationRequest:
      type: object
      description: Request to create a product reservation
      properties:
        productId:
          type: string
          format: uuid
          description: UUID of the product to reserve
        userId:
          type: string
          format: uuid
          description: UUID of the user making the reservation
        quantity:
          type: integer
          format: int32
          description: Quantity of items to reserve
          example: 5
    ReservationResponse:
      type: object
      description: Reservation response payload
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the reservation
        productId:
          type: string
          format: uuid
          description: ID of the reserved product
        userId:
          type: string
          format: uuid
          description: ID of the user who made the reservation
        status:
          type: string
          description: Current status of the reservation
          enum:
          - PENDING
          - CONFIRMED
          - EXPIRED
          - CANCELLED
        quantity:
          type: integer
          format: int32
          description: Quantity of items reserved
        createdAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the reservation was created
        updatedAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the reservation was last updated
        expiresAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the reservation will expire
    CreateProductRequest:
      type: object
      description: Request to create a new product
      properties:
        name:
          type: string
          description: Name of the product
          example: Smartphone
        stock:
          type: integer
          format: int32
          description: Initial stock quantity
          example: 100
    ProductResponse:
      type: object
      description: Product response payload
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the product
        name:
          type: string
          description: Name of the product
          example: Laptop
        stock:
          type: integer
          format: int32
          description: Current stock quantity
          example: 50
        version:
          type: integer
          format: int64
          description: Optimistic locking version
        createdAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the product was created
        updatedAt:
          type:
          - string
          - "null"
          format: date-time
          description: Timestamp when the product was last updated